requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.20.0
pydantic>=1.10.0
pytest>=7.0.0
httpx>=0.24.0
sqlalchemy>=2.0.0
python-dotenv>=0.21.0
pytest-asyncio>=0.21.0
